+++ octoprint_astroprint/static/js/astroprint.js
@@ -27,6 +27,7 @@ $(function () {
         self.description = ko.observable("");
         self.access_key = ko.observable("");
         self.boxName = ko.observable(astroprint_variables.boxName);
+        self.boxId = ko.observable(astroprint_variables.boxId);
         self.printerModel = ko.observable(astroprint_variables.printerModel)
         self.cacheBoxName = ko.observable(self.boxName());
 
@@ -702,6 +703,7 @@ $(function () {
               currentUrl = encodeURI(currentUrl);
               var url = astroprint_variables.appSite + "/authorize" +
                   "?client_id=" + astroprint_variables.appId +
+                  "&box_id=" +
                   "&redirect_uri=" + currentUrl +
                   "&scope=" + encodeURI("profile:read project:read design:read design:download print-file:read print-file:download print-job:read device:connect device:update")+
                   "&state="+ap_access_key+
