+++ octoprint_astroprint/templates/astroprint_tab.jinja2
@@ -3,6 +3,7 @@
     appSite : '{{ plugin_astroprint_appSite|escape }}',
     appId : '{{ plugin_astroprint_appId|escape }}',
     boxName : '{{plugin_astroprint_boxName|escape}}',
+    boxId : '{{plugin_astroprint_boxId|escape}}',
     printerModel : {{plugin_astroprint_printerModel}},
     filament : {{plugin_astroprint_filament}}
   }
